refactor: extract thread display state
